Basic game play functionality complete

+++ Columbo.Minesweeper.Application/Commands/FinishGameCommandHandler.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using Columbo.Minesweeper.Application.Commands.Infrastructure;
 
 namespace Columbo.Minesweeper.Application.Commands
 {
@@ -18,7 +19,8 @@ namespace Columbo.Minesweeper.Application.Commands
         {
             var game = _game_repository.get_by(command.player_id);
 
-            _game_repository.delete(game);
+            if (game != null)
+                _game_repository.delete(game);
         }
     }
 }
